Add netcoreapp2.0 config to SystemEvents
Adding a netcoreapp2.0 configuration to SystemEvents as the source
assembly targets netcoreapp2.0 and this avoid the netstandard shim to
be used. The additional configuration shouldn't pose a problem as with
an eventual convergance of reference and src projects, we will likely
not mix universes between ref and src anymore and build the additional
matching universe assemblies.