Simplify sequential layout check per JanK's suggestion
I have simplified runtime and Crossgen2 behavior in the presence
of sequential layout by removing blittability from the picture
as the IsBlittable check is very complex and hard to maintain in
sync between the native runtime and the managed compiler. I have
introduced a somewhat weaker and much simpler check
MayContainGCPointers that returns FALSE for all IsBlittable types
(and for a few others).

Thanks

Tomas