Fixes for x86 corner cases including a pre-existing blittability bug
In my investigation of the x86 failures I hit an interesting corner
case of the class GCMemoryInfoData that is checked for having managed
layout matching its native counterpart. Interestingly enough, that
only worked because the class internally reported as non-blittable
and not managed sequential so it ended up using the auto layout;
when my change switched it over to use the sequential layout, it
turned out that the sequential layout algorithm differs from its
native counterpart. While the change is technically breaking this
corner scenario, I suspect that the primary purpose of sequential
layout is interop with native code and I find the fact that it's
actuall mismatched concerning beyond the scope of my pending change.

Thanks

Tomas