Add additional .NETCoreApp configurations to avoid NS shim
Adding netcoreapp3.1 assets to libraries to avoid the
netstandard.dll shim as a dependency. The secondary motive
for adding these configurations is to keep source and ref
configurations in sync to avoid weird cases where P2Ps
resolve assets from different assembly universes depending
on if it's a source or reference assembly.

The only libraries that still go through the netstandard.dll
shim are the Microsoft.Extensions ones.