Remove a failing unnecessary test
The manually assigned 128 KiB stack here is bordering on the amount of
stack space that release JIT needs to JIT functions with deep trees.
Since the JIT already has a limit on the depth of trees it creates

https://github.com/dotnet/runtime/blob/44f050acc0f7791d6cd7ac772945444912bcf299/src/coreclr/jit/importer.cpp#L11234-L11252

it seems unnecessary to pressimize the JIT further to allow this test to
pass. Also, the test was originally added as a
faster equivalent to the test above it that could run in inner loop
(https://github.com/dotnet/corefx/pull/14444), but it was subsequently
moved to outer loop (https://github.com/dotnet/corefx/pull/19563), so
now it does not seem like there's much point in retaining it.

Fix #53309