Fix the issue that with regions we are unable to allocate objects larger than 32 MB.
Method gc_heap::get_segment_for_uoh gets passed a size, but drops it on the floor in the region case.

The fix is simply to pass the size parameter through the various methods we call for allocating a large region, and to allocate a properly sized region in region_allocator::allocate_large_region.