Consolidate SSE shuffle constant unrolling
Remove OP_SSE2_SHUFFLE, which is unused

Rename OP_SSE_SHUFFLE to OP_SSE_SHUFPS, to make this more consistent
with the naming convention used for other SSE shuffles

Use `immediate_unroll_*` instead of hand-writing branch emission. These
branch tables are huge (in the simplest case, with 256 different
constant values, we can spend over 1KB of code on nothing but `shufps`
and `jmp`s, and the cost gets worse if any tail duplication happens),
and are currently emitted inline. Future work ought to:

1. use a sequence of extractelement/insertelement instructions, which
can be optimized into a constant shuffle when the shuffle control
parameter is constant, and otherwise generates a high-latency but
low-code-size fallback (note that this only works for shuffles); or

2. emit the fallback branch tables out of line and use
`llvm.is.constant` to generate either a constant shuffle or a call to a
fallback shuffle branch table function (the cost isn't too bad: a
direct-call/ret pair would add ~4-5 cycles and eat an RSB slot on top of
the cost of the branch table).

Fixes `JIT/HardwareIntrinsics/X86/Regression/GitHub_21855/GitHub_21855_r`