Be more verbose when processes spawned from tasks fail.
Right now, when a process fails (e.g. cmake or xcodebuild), all of the
STDERR captured during that process run is dumped in the exception.

However, if silent is true, no STDOUT is captured at all, and we lose
potentially vital debugging information contained within.

This PR merges the stdout and stderr capture, and instead of not
capturing anything when silent is true, it always captures and only
returns nothing when silent is true.

This should greatly ease debugging of failed tasks.
a program's output on error