This fixes an issue in Server GC where an item in the finalizer queue became stale due to not being relocated.
The problem was that a finalizable object was allocated on one heap, but registered in the finalizer queue of another heap (this is possible due to heap balancing). In CFinalize::UpdatePromotedGenerations, we ask for the generation of an object, and move the object to the correct section of the finalizer queue. In the error case, we obtained the wrong result for the generation of the object because it lives on another heap, and that heap hadn't set the final generation for the region containing the object yet.

The fix simply fetches the heap from the alloc context after the allocation. That is the correct heap to register the object on.