Pass the default ALC gchandle to the v3 preload hook
Fixes crashes early during startup when the gchandle is then used by the hooks
to call mono API functions.  The default ALC exists, but its managed object
doesn't, so the gchandle target is null.  The mono APIs detect the the special
gchandle and resolve it to the default ALC

Example crash
https://gist.github.com/grendello/b4ab24587a055725cc5e1416b86ad7ca