Fix BackoutJitData
The RemoveJitData that the BackoutJitData calls requires the code
header to be copied to the final location. This change fixes it.

I've also found that in one of my previous changes, I've accidentally
enabled jitting into a scratch buffer by default by adding the
FEATURE_WXORX define. So I am removing it in this change, it will
be replaced by a dynamic check whether W^X is enabled in the final
W^X change.