Optimize `CAST(int <- long)` on 32 bit targets (#53040)
* Optimize CAST(int <- long) on 32 bit targets

* Revert "Optimize CAST(int <- long) on 32 bit targets"

Revert the implementation in lowering

* Optimize CAST(int <- long) on 32 bit targets

Move the code from lowering to long decomposition.

* Fixed the "Arguments" note for DecomposeNode

* Added the function header

For OptimizeCastFromDecomposedLong.

* Remove the TODO comment

While correct, it has questionable value.

* Add a more detailed dump output

* Do not try to optimize checked casts

It is easy to get it wrong. Let the frontend handle this.

* Do not depend on tree order

Previous version of the code assumed that there could be
no nodes between the cast and its operand. That is not
a correct assumption to make in LIR.