[metadata_update] Return "Baseline" capabilities even if env is unset
MetadataUpdater.GetCapabilities() on CoreCLR always returns the supported
edits, even if in the current running instance the environment variables for
hot reload or EnC are unset.

This change makes mono behave the same, except in the case where the runtime is
built without hot reload support at all.

MetadataUpdater.IsSupported still behaves the same as before at runtime - it
returns true if there is some possibility that this ApplyUpdate could be called
and could succeed (ie env is set apropriately)