Fix ordinal imports on Windows.
Fixes mono/mono#21122

We have existing code for this, but as the test case shows, it doesn't work. It's not really clear how it's supposed to work, since all it does is skip the "#" in the entry point name. GetProcAddress does not interpret numeric strings as ordinals, but it will interpret a pointer address from 1 to 65535 as an ordinal.