Noticed we ran out of memory with regions on the GC stress test.
The fix is to limit the number of regions that go on the per-heap free lists.

Whether a region should be returned to the region allocator is decided by a new API region_allocator::should_delete_region.

The current implementation takes a couple things into consideration, subject to future tuning:
- the address space that is tied up in free regions per heap should not get too large
- when the va memory load exceeds 50%, we want to get more stingy with virtual address space.
- we want to try to keep the upper half of the address space less fragmented

The changes also include some improvements to dprintf calls.