FileSystem.Exists.Unix: eliminate a stat call when checking a non-existing file/directory.
When checking for a directory, we can use `stat`, there is no need to make an `lstat` call.

When checking for a file, use `lstat` which will give us what we need in most cases.
Only when it is a link, we need to call `stat` to check if the final target is not a directory.